fixing errors...
when the program is run, only around 89% of ducks are generated... I don't know why.
right now I'm trying to fix it by removing the txt file to array function outside the main loop. if this doesn't fix it, the program will still be more efficient because it wont turn the file into an array each time anymore.
I tried to fix the error with a time gap between each iteration but the error was still there...